bug fix using /playerinfo on an OfflinePlayer, removed unnessary code from CanaryServer constructor

+++ src/main/java/net/canarymod/api/CanaryOfflinePlayer.java
@@ -192,7 +192,7 @@ public class CanaryOfflinePlayer implements OfflinePlayer {
     @Override
     public long getTimePlayed() {
         if (getNBT() != null && getNBT().containsKey("TimePlayed")) {
-            return getNBT().getInt("TimePlayed");
+            return getNBT().getLong("TimePlayed");
         }
         return 0;
     }
